// src/usecase/customer/update/update.customer.dto.ts
export interface InputUpdateustomerDTO {
  id: string
  name: string
  address: {
    street: string
    number: number
    city: string
    zip: string
  }
}

export interface OutputUpdateCustomerDTO {
  id: string
  name: string
  address: {
    street: string
    number: number
    city: string
    zip: string
  }
}